Conflicts:
	doc/Section_errors.txt

+++ src/USER-EFF/atom_vec_electron.h
@@ -57,7 +57,7 @@ class AtomVecElectron : public AtomVec {
   int data_atom_hybrid(int, char **);
   void data_vel(int, char **);
   int data_vel_hybrid(int, char **);
-  double memory_usage();
+  bigint memory_usage();
   
  private:
   double PI;
@@ -71,4 +71,3 @@ class AtomVecElectron : public AtomVec {
 
 #endif
 #endif
-
