+++ src/update.h
@@ -15,7 +15,6 @@
 #define LMP_UPDATE_H
 
 #include "pointers.h"
-#include "lmptype.h"
 
 namespace LAMMPS_NS {
 
@@ -52,7 +51,7 @@ class Update : protected Pointers {
   void create_integrate(int, char **);
   void create_minimize(int, char **);
   void reset_timestep(int, char **);
-  double memory_usage();
+  bigint memory_usage();
 };
 
 }
